Move playwright tests into integration test folder.
