+++ integration-test/data/har/duckduckgo.com/extension-success.har
