+++ integration-test/data/staticcdn/https/https-allowlist.json
