+++ integration-test/data/staticcdn/https/https-bloom.json
