+++ integration-test/data/staticcdn/https/negative-https-allowlist.json
