+++ integration-test/data/staticcdn/https/negative-https-bloom.json
