+++ integration-test/data/staticcdn/trackerblocking/config/v2/extension-chrome-config.json
